Use Optional.ofNullable instead of Optional.of
Since links don't have to have names, we should use Optional.ofNullable
whose parameter can be null, as opposed to Optional.of whose parameter
cannot be. If we don't do this, we get some NPEs since the original
links in the tests are nameless.

NB: This fixes 2 of 5 test failures.