Add test to prevent linking items of differing types
Add a test that ensures we don't allow linking a source and target of differing
data types.

Note: Unfortunately, this is not working as expected... In the Eclipse debugger
I have access to `linkSource.tail.type.name` but I don't have the ability of
using this in any code, it seems like. I still need to figure this out but, by
this point, I'm unsure as to how to proceed.