Implement test to prevent refining data
Implement the test that prevents a user from trying to refine a data type. This
test is potentially missing something but, as of right now, I'm not sure as to
what that might be. There doesn't seem to be a straightforward way of checking
if the model that was passed in is, in fact, a model. Thus, what I'm doing here
is checking to see if there is a refined model but the model has no name. This
indicates that the model failed to load and, as such, is not a model after all.