Clean up RefinedModelValidator
Since the same function that prevents users from refining data types also seems
to prevent circular refinements, remove the function specifically dedicated to
the latter and rename the former to something that can apply to both cases.

Add a comment (AHH!) to warn of why we have to do what we're doing.