Start refactoring refinedModelHasValidRequiresBlock()
The refinedModelHasValidRequiresBlock is too big so start refactoring it.
Initially, we get rid of the specific calls to get the requirements for each of
the models and replace them with the necessary calls to get requirement
declarations for each of the models instead since the other step was really only
used to see if we should try to get the requirement declarations.