Add a valid requires refinement to the test case
To confirm that we should be allowed to refine a requirement that is contained
in the model that is being refined, add a reference to the requirement contained
in the Alarm model. Eventually, this will need to be updated to include the
'refines' keyword - whenever that functionality is implemented, that is.