Add initial property enumeration value test and implementation
I still need to figure out how to confirm that the value is set. I suspect the
failure might be due to the fact that I'm using a mock so nothing ends up
getting set, but I still need to confirm if that's actually the reason the test
is failing.